♻️ refactor: #115 마커별 스토리에서 이미지 컴포넌트 분리
