+++ src/components/map/MapDetailStory.tsx
@@ -1,15 +1,10 @@
 'use client';
 
-import Image from 'next/image';
 import { useRouter } from 'next/navigation';
-import { useQueries, useQuery } from '@tanstack/react-query';
 
-import { markerOption } from '@/api/options/markerOption';
-import { storyOption } from '@/api/options/storyOption';
-import { gridLayout } from '@/components/map/map.recipe';
+import MapDetailStoryImages from '@/components/map/MapDetailStoryImages';
 import { flex, flexBetween } from '@/components/ui/common/cards/card.recipe';
 import { modalText } from '@/components/ui/common/modals/modal.recipe';
-import { Story, StoryImage } from '@/types/story';
 
 type MapDetailStoryProps = {
   id: number;
@@ -18,12 +13,6 @@ type MapDetailStoryProps = {
 
 function MapDetailStory({ id, title }: MapDetailStoryProps) {
   const router = useRouter();
-
-  const { data: storyList } = useQuery(markerOption.getMarkerStory(id));
-  const storyIds = storyList?.data?.map((story: Story) => story.storyId) ?? [];
-  const storyImageQueries = useQueries({
-    queries: storyIds.map(storyId => storyOption.getStoryImage(storyId)),
-  });
 
   return (
     <div className={flex({ gap: 'md' })}>
@@ -50,25 +39,7 @@ function MapDetailStory({ id, title }: MapDetailStoryProps) {
           더보기
         </span>
       </p>
-      <div className={gridLayout({ columns: 4, gap: 'sm', p: 'xs' })}>
-        {storyImageQueries.map((query, i) => {
-          const images = query.data?.data ?? [];
-          return images.map((img: StoryImage, j) => (
-            <Image
-              key={`${i}-${j}`}
-              src={img.imageUrl}
-              alt={title}
-              width={0}
-              height={0}
-              style={{
-                width: '100%',
-                height: 'auto',
-                borderRadius: 8,
-              }}
-            />
-          ));
-        })}
-      </div>
+      <MapDetailStoryImages id={id} title={title} />
     </div>
   );
 }
